+++ abkit/stats/parametric/paired_ttest.py
@@ -111,7 +111,8 @@ class PairedTTest(BasePairedMethod):
             cov_value_1 = float(moments.mean[moments.index("x1")])
             cov_value_2 = float(moments.mean[moments.index("x2")])
 
-        return TestResult(
+        return self._result_from_normal_test(
+            test,
             name_1=joint.name_1,
             name_2=joint.name_2,
             value_1=float(moments.mean[index_y1]),
@@ -122,15 +123,4 @@ class PairedTTest(BasePairedMethod):
             size_2=n,
             cov_value_1=cov_value_1,
             cov_value_2=cov_value_2,
-            method_name=self.name,
-            method_params=self.identity_params,
-            alpha=self.alpha,
-            pvalue=test.pvalue,
-            effect=test.effect,
-            ci_length=test.ci_length,
-            left_bound=test.left_bound,
-            right_bound=test.right_bound,
-            reject=test.reject,
-            effect_distribution=test.distribution,
-            warnings=test.warnings,
         )