+++ abkit/stats/parametric/ttest.py
@@ -78,7 +78,8 @@ class TTest(BaseMethod):
                 ratio=stats_1.n / stats_2.n,
             )
 
-        return TestResult(
+        return self._result_from_normal_test(
+            test,
             name_1=stats_1.name,
             name_2=stats_2.name,
             value_1=stats_1.mean,
@@ -91,15 +92,4 @@ class TTest(BaseMethod):
             cov_value_2=stats_2.cov_mean,
             mde_1=mde_1,
             mde_2=mde_2,
-            method_name=self.name,
-            method_params=self.identity_params,
-            alpha=self.alpha,
-            pvalue=test.pvalue,
-            effect=test.effect,
-            ci_length=test.ci_length,
-            left_bound=test.left_bound,
-            right_bound=test.right_bound,
-            reject=test.reject,
-            effect_distribution=test.distribution,
-            warnings=test.warnings,
         )